Add a filter for area selection
You can now pick your area in the start menu and it will update the map.
It is not allowed to add more then one structural angular directive in
a tag. This is why I added `<ng-container>` and divide it over two tags.
You can repeat this as much as you want